cloud: fixup for #404
