+++ dvc/cloud/instance_manager.py
@@ -1,0 +1,4 @@
+class CloudSettings(object):
+    def __init__(self, global_storage_path, cloud_config):
+        self.cloud_config = cloud_config
+        self.global_storage_path = global_storage_path
